+++ app/routes/__auth/activate.jsx
@@ -8,7 +8,8 @@ import {
 import { authenticator } from '~/auth/auth.server';
 import { sessionStorage } from '~/auth/session.server';
 import { AuthorizationError } from 'remix-auth';
-import { CmsAuth, CmsError } from '~/api/cms.server';
+import { CmsAuth } from '~/api/cmsauth.server';
+import { CmsError } from '~/api/util';
 
 export function CatchBoundary() {
 	const caught = useCatch();
